+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # CHANGELOG
 
+## Bugfree.Spo.Analytics 1.2.1 (2017-04-22)
+
+* About 1% of x-forwarded-for headers fail to parse due to change in header format ([#6](https://github.com/ronnieholm/Bugfree.Spo.Analytics/issues/6))
+
 ## Bugfree.Spo.Analytics 1.2.0 (2017-04-17)
 
 * Updated documentation with build prerequisites
 * Updated build script to use LocalDb 2016 and Visual Studio 2017 Professional
-* Implement more complete Url parser for better visit Url inspection and filtering ([#5](https://github.com/ronnieholm/Bugfree.Spo.Analytics/issues/5))
+* Implemented more complete Url parser for better visit Url inspection and filtering ([#5](https://github.com/ronnieholm/Bugfree.Spo.Analytics/issues/5))
 
 ## Bugfree.Spo.Analytics 1.1.0 (2016-11-06)